+++ lib/dropdown_search/properties/dropdown_decorator_props.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class HcDropDownDecoratorProps {
+  final InputDecoration? dropdownSearchDecoration;
+  final TextStyle? baseStyle;
+  final TextAlign? textAlign;
+  final TextAlignVertical? textAlignVertical;
+
+  const HcDropDownDecoratorProps({
+    this.dropdownSearchDecoration,
+    this.baseStyle,
+    this.textAlign,
+    this.textAlignVertical,
+  });
+}